update test images to use newest pinned golang
