+++ internal/ui/etui_event_handlers.go
